# backend/docs/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/petclinic/v1/owner":{"get":{"tags":["owner-controller"],"summary":"find owner by telephone","description":"find an owner by the telephone","operationId":"findOwner","parameters":[{"name":"telephone","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OwnerResponse"}}}}}},"post":{"tags":["owner-controller"],"summary":"create an owner","description":"Creates an owner in the platform if the owner does not exist. The requiredunique key is a telephone. No two owners will have the same phonenumber.","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOwnerRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OwnerResponse"}}}}}}},"/petclinic/v1/owners":{"get":{"tags":["owner-controller"],"summary":"find all owners","description":"returns all owners that are available in the system","operationId":"findOwners","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OwnerResponse"}}}}}}},"post":{"tags":["owner-controller"],"summary":"create an owner","description":"Creates a list of owners in the platform if the owner does not exist. The requiredunique key is a telephone. No two owners will have the same phonenumber.","operationId":"bulkCreate","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CreateOwnerRequest"}}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OwnerResponse"}}}}}}}},"/petclinic/v1/owner/{id}":{"get":{"tags":["owner-controller"],"summary":"find owner by id","description":"find owner by id","operationId":"findOwnerById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"include_pets","in":"query","required":false,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OwnerResponse"}}}}}}},"/petclinic/v1/owner/{id}/pets":{"get":{"tags":["pet-controller"],"summary":"find all pets for an owner","operationId":"findPetsByOwner","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PetResponse"}}}}}}}},"/petclinic/v1/owner/{id}/pet":{"post":{"tags":["pet-controller"],"summary":"create a new pet for given the owner's id","operationId":"createPet","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PetRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PetResponse"}}}}}}},"/petclinic/v1/owner/pet":{"post":{"tags":["pet-controller"],"description":"create a new pet. Also create the owner if it does not exist.","operationId":"createPetWithOwner","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PetRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PetResponse"}}}}}}},"/petclinic/v1/specialties":{"get":{"tags":["specialty-controller"],"summary":"find all specialties","description":"find all available specialties available.","operationId":"findAll","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SpecialtyResponse"}}}}}}}},"/petclinic/v1/specialty":{"post":{"tags":["specialty-controller"],"summary":"create a specialty","description":"create a specialty if does not exist.","operationId":"createSpecialty","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpecialtyRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpecialtyResponse"}}}}}}},"/petclinic/v1/vets":{"get":{"tags":["vet-controller"],"summary":"find all vets","description":"find all vets available without any filters","operationId":"findAll_1","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VetResponse"}}}}}}}},"/petclinic/v1/vets/{id}":{"get":{"tags":["vet-controller"],"summary":"find a vet by id","description":"find vet by id.","operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VetResponse"}}}}}}},"/petclinic/v1/vets/{id}/specialty":{"put":{"tags":["vet-controller"],"summary":"add a specialty to a vet","description":"add a specialty to vet if the vet does not have the specialty","operationId":"addSpecialtyToVet","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpecialtyRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VetResponse"}}}}}}},"/petclinic/v1/vet":{"post":{"tags":["vet-controller"],"summary":"create a vet","description":"create a new vet in the system","operationId":"createVet","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateVetRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VetResponse"}}}}}}},"/petclinic/v1/visits":{"get":{"tags":["visit-controller"],"operationId":"findVisits","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Visit"}}}}}}},"post":{"tags":["visit-controller"],"operationId":"createVisit","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VisitRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateVisitResponse"}}}}}}}},"components":{"schemas":{"CreateOwnerRequest":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"first_name":{"type":"string"},"last_name":{"type":"string"},"telephone":{"type":"string"},"address":{"type":"string"},"city":{"type":"string"}}},"OwnerResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"first_name":{"type":"string"},"last_name":{"type":"string"},"telephone":{"type":"string"},"address":{"type":"string"},"city":{"type":"string"},"pets":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/PetResponse"}}}},"PetResponse":{"type":"object","properties":{"id":{"type":"string"},"owner_id":{"type":"string"},"birth_date":{"type":"string","format":"date"},"type":{"type":"string"},"name":{"type":"string"},"visits":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Visit"}}}},"Visit":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"date":{"type":"string","format":"date-time"},"petId":{"type":"string","format":"uuid"},"ownerId":{"type":"string","format":"uuid"},"vetId":{"type":"string","format":"uuid"},"description":{"type":"string"}}},"PetOwnerRequest":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"telephone":{"type":"string"}}},"PetRequest":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"owner_id":{"type":"string","format":"uuid"},"birth_date":{"type":"string","format":"date"},"type":{"$ref":"#/components/schemas/PetType"},"name":{"type":"string"},"owner":{"$ref":"#/components/schemas/PetOwnerRequest"},"visits":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Visit"}}}},"PetType":{"type":"object","properties":{"name":{"type":"string"}}},"SpecialtyResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"}}},"SpecialtyRequest":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"}}},"VetResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"first_name":{"type":"string"},"last_name":{"type":"string"},"specialties":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/SpecialtyResponse"}}}},"CreateVetRequest":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"first_name":{"type":"string"},"last_name":{"type":"string"},"telephone":{"type":"string"},"email":{"type":"string"},"specialties":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}},"VisitRequest":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"date":{"type":"string","format":"date"},"pet":{"type":"string","format":"uuid"},"owner":{"type":"string","format":"uuid"},"vet":{"type":"string","format":"uuid"},"description":{"type":"string"}}},"CreateVisitResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"date":{"type":"string","format":"date-time"},"pet":{"type":"string","format":"uuid"},"owner":{"type":"string","format":"uuid"},"vet":{"type":"string","format":"uuid"},"description":{"type":"string"}}}}}}